Add new is_valid test for Id_Task & Tests & move models to a models file
