+++ ngtq_bus_service_client/src/lib.rs
@@ -1,45 +1,7 @@
 use std::{io::{Read, Write}, os::unix::net::UnixStream};
-use serde::{Deserialize, Serialize};
+use models::{BusRequest, CategoryTask, IdTask, Task};
 
-#[derive(Debug, Serialize, Deserialize)]
-enum BusRequest {
-    PushTask(Task),
-    PullTask(TaskIdentifier),
-    Error(String)
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-enum TaskIdentifier {
-    Id(String),
-    Category(String),
-    Error(String)
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct IdTask {
-    id: String,
-    payload: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct CategoryTask {
-    id: String,
-    payload: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub enum Task {
-    Id(IdTask),
-    Category(CategoryTask),
-    Error(String)
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct BusResponse {
-    successful: bool, 
-    error: String, 
-    payload: String
-}
+mod models;
 
 pub struct BusServiceClient {
     is_initialise: bool,
@@ -56,15 +18,21 @@ impl BusServiceClient {
             return Err(String::from("Failed to send task - BusServiceClient is not initilised!"))
         }
 
-        let serialise_task = serde_json::to_string(&BusRequest::PushTask(Task::Id(id_task))).unwrap();
-        match send_request_to_bus(serialise_task, &self.bus_address) {
-            Ok(response) => {
-                match response.parse::<usize>() {
-                    Ok(queue_size) => Ok(queue_size),
-                    Err(error) => Err(format!("Failed to parse queue size to usize: {}", error.to_string()))
+        match id_task.is_valid() {
+            Ok(_) => {
+
+                let serialise_task = serde_json::to_string(&BusRequest::PushTask(Task::Id(id_task))).unwrap();
+                match send_request_to_bus(serialise_task, &self.bus_address) {
+                    Ok(response) => {
+                        match response.parse::<usize>() {
+                            Ok(queue_size) => Ok(queue_size),
+                            Err(error) => Err(format!("Failed to parse queue size to usize: {}", error.to_string()))
+                        }
+                    },
+                    Err(error) => Err(error)
                 }
             },
-            Err(error) => Err(error)
+            Err(error) => return Err(error)
         }
     }
